backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
google-generativeai==0.8.3
reportlab==4.2.0
Pillow==10.4.0
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
jinja2==3.1.4
